Main-Site Style Changes
Adjusted the poll title to scale with the change in window size and various improvments to the buttons on the vote page of the poll. I tried to implement "FitText" jquery library but failed to get it to work due to the elements that need scaling are hidden when the page loads.